add help tab to lettering
