+++ lib/gui/lettering/__init__.py
@@ -1,0 +1,8 @@
+# Authors: see git history
+#
+# Copyright (c) 2024 Authors
+# Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
+
+from .help_panel import LetteringHelpPanel
+from .option_panel import LetteringOptionsPanel
+from .main_panel import LetteringPanel
