+++ lib/gui/satin_multicolor/__init__.py
@@ -1,0 +1,8 @@
+# Authors: see git history
+#
+# Copyright (c) 2024 Authors
+# Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
+
+from .colorize_panel import ColorizePanel
+from .help_panel import HelpPanel
+from .main_panel import MultiColorSatinPanel
