+++ lib/gui/satin_multicolor/colorize_panel.py
